t264: Remove unused properties from GPC DMA
Remove nvidia,start-dma-channel-index and dma-channels. They were
added for backward compatibility with k-5.10. Current driver
does not use these properties anymore.

Remove iommus  and dma-coherent since GPCDMA driver is not allocating
any memory on its own in T264. The memory is expected to be owned
by the respective clients and passed on to GPCDMA driver. This also
fixes a warning in the hypervisor for AV+L.

Bug 4709342

Change-Id: I8d99af9d1865e9a81d22d4b35440d4c2eeebd5fe
Reviewed-on: https://git-master.nvidia.com/r/c/device/hardware/nvidia/t264-public-dts/+/3160914
Tested-by: mobile promotions <svcmobile_promotions@nvidia.com>
Reviewed-by: mobile promotions <svcmobile_promotions@nvidia.com>