t23x: nv-public: Disable hot surface alert for IGX
For IGX platform, the module and baseboard are inside the chassis, the
hot surface alert should not rely on on-chip or on-board thermal sensors
to judge whether the surface is hot or not. So, this change disables the
hot surface alert for IGX platform to avoid the false alarm.

Bug 4084478
Bug 4561083
Bug 4611631

Signed-off-by: Yi-Wei Wang <yiweiw@nvidia.com>
Change-Id: Ib4f9ca3e2822c593744668a82ebaabb3910ac594
Reviewed-on: https://git-master.nvidia.com/r/c/device/hardware/nvidia/t23x-public-dts/+/3181470
(cherry picked from commit 82746253aa02cf3c0d99fa1404a906ec6f55bb44)
Reviewed-on: https://git-master.nvidia.com/r/c/device/hardware/nvidia/t23x-public-dts/+/3182331
GVS: buildbot_gerritrpt <buildbot_gerritrpt@nvidia.com>
Reviewed-by: Rajkumar Kasirajan <rkasirajan@nvidia.com>