dts: t264: remove unused psci-domain-idle driver
Removed definitions and structure required for the psci-domain-idle
driver.
This was created to support CR7, however, the CR7-idle feature has been
dropped from t264.
As CR7 is never enabled, this is causing a kmemleak during boot that
needs to be fixed.
The driver and structure is not needed, so it can be replaced with the
simpler psci-cpuidle driver which does not contain memory leaks.

Bug 4746881
Bug 4816091

Change-Id: I0c84a59b210db07b8d3224fd78481dc1a90c6713
Reviewed-on: https://git-master.nvidia.com/r/c/device/hardware/nvidia/t264-public-dts/+/3198045
Tested-by: mobile promotions <svcmobile_promotions@nvidia.com>
Reviewed-by: mobile promotions <svcmobile_promotions@nvidia.com>