NVIDIA: SAUCE: drivers: nv_ist: Add IST debugfs options
Adding a driver to handle setup and trigger for IST through the L4T
kernel.
This driver "hijacks" the current Hotplug entry/exit flow to instead
trigger the core to enter the RIST state.
Upon failure or success, the core will then be left in the "unplugged"
state, to be brought back online at a later time manually.
This set of changes is mainly meant for testing RIST in L4T-based
systems. This is not intended for production use-cases.

TPS-589

Signed-off-by: Ishan Shah <ishah@nvidia.com>
Signed-off-by: Laxman Dewangan <ldewangan@nvidia.com>
Acked-by: Noah Wager <noah.wager@canonical.com>
Acked-by: Jacob Martin <jacob.martin@canonical.com>
Signed-off-by: Noah Wager <noah.wager@canonical.com>