nv-public: modularize thermal features
Thermal features are highly dependent on both the software use cases
and the hardware thermal solutions. For example, thermal slowdown is
not required for the safety platforms; shutdown thresholds can be higher
with the industrial grade components, and so on.

With modularized thermal device tree files, it will be easier to
maintain platform-specific thermal settings.

Bug 4459419

Signed-off-by: Yi-Wei Wang <yiweiw@nvidia.com>
Change-Id: I189a05980d07bffbce9493c56d333c3145ca3d0c
Reviewed-on: https://git-master.nvidia.com/r/c/device/hardware/nvidia/t264-public-dts/+/3224852
Reviewed-by: Brad Griffis <bgriffis@nvidia.com>
Reviewed-by: Rajkumar Kasirajan <rkasirajan@nvidia.com>
GVS: buildbot_gerritrpt <buildbot_gerritrpt@nvidia.com>