NVIDIA: SAUCE: net: phy: aquantia: Poll for valid FW ID.
Ethernet link up fails after SC7 resume as ethernet settings are
configured before phy firmware is completely loaded.

Firmware ID will be shown as 0xFFFF if firmware is not yet loaded.
So, add this too in the exit condition of the poll so that poll
won't exit for 0xFFFF and continues till init completion with valid
FW ID.

http://nvbugs/4382502
http://nvbugs/4980644

Signed-off-by: Revanth Kumar Uppala <ruppala@nvidia.com>
Reviewed-by: Bhadram Varka <vbhadram@nvidia.com>
Reviewed-by: Pohsun Su <pohsuns@nvidia.com>
Reviewed-by: Narayan Reddy <narayanr@nvidia.com>
Reviewed-by: Bibek Basu <bbasu@nvidia.com>
Reviewed-by: Jon Hunter <jonathanh@nvidia.com>
Signed-off-by: Vishwaroop A <va@nvidia.com>
Acked-by: Noah Wager <noah.wager@canonical.com>
Acked-by: Jacob Martin <jacob.martin@canonical.com>
Signed-off-by: Noah Wager <noah.wager@canonical.com>