nv-public: Enable dynamic fan speed ctrl for p4071
This change adds preliminary fan settings for p3834-0008 + p4071
platform that has only one fan installed on top of the module. Hence,
this change also removes the corresponding pwm-fan-nvme node as no
dedicated fan is installed to cool down the nvme.

On the other hand, the fan speed to temperature mapping should be
implemented in the platform-specific base dts as the settings vary among
the combinations of the module skus and the carrier boards, so
restructure them accordingly.

Bug 4459411

Signed-off-by: Yi-Wei Wang <yiweiw@nvidia.com>
Change-Id: Id7321d53d7636f6a4a174f34d771199d4ebf6ba8
Reviewed-on: https://git-master.nvidia.com/r/c/device/hardware/nvidia/t264-public-dts/+/3314710
GVS: buildbot_gerritrpt <buildbot_gerritrpt@nvidia.com>
Tested-by: mobile promotions <svcmobile_promotions@nvidia.com>
Reviewed-by: mobile promotions <svcmobile_promotions@nvidia.com>
Reviewed-by: Rajkumar Kasirajan <rkasirajan@nvidia.com>