NVIDIA: SAUCE: phy: tegra: fix port reset issue
XHCI driver will dynamically power off UTMI pad for root hub ports
entering suspend for power-saving.

If the device triggers a remote wakeup, XHCI driver needs to respond
the resume signal within 1ms according to USB2 spec. This means UTMI
pad power should be powered on within 1ms. However, when USB core driver
receives a "port status change" event for RESUME, it will wait 40ms
to drive the resume signaling on the bus. During this period of time,
XHCI driver has no chance to power on the UTMI pad which leads to the
disconnection of USB device.

In this change, we decide to remove the dynamic powerdown of UTMI pad
for autosuspend.

http://nvbugs/4819335
http://nvbugs/4752661
http://nvbugs/3370448
http://nvbugs/5181436

Signed-off-by: EJ Hsu <ejh@nvidia.com>
Reviewed-by: Henry Lin <henryl@nvidia.com>
Reviewed-by: HaoTien Hsu <haotienh@nvidia.com>
Reviewed-by: WK Tsai <wtsai@nvidia.com>
Signed-off-by: Vishwaroop A <va@nvidia.com>
Acked-by: Noah Wager <noah.wager@canonical.com>
Acked-by: Jacob Martin <jacob.martin@canonical.com>
Signed-off-by: Noah Wager <noah.wager@canonical.com>